RPC and block device classes
